requirements.txt
transformers>=4.35.0
torch>=2.0.0
datasets
accelerate

openai-whisper
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0

gigachat
langchain>=0.1.0
langchain-gigachat

aiogram>=3.3.0
aiohttp>=3.9.0

pandas>=2.0.0
scikit-learn>=1.3.0

python-dotenv
pydantic>=2.0.0
spotipy>=2.23.0
